activity-service/requirements.txt
grpcio
grpcio-tools
grpcio-health-checking
flask
redis
pymongo
